+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@pchmn/expo-material3-theme": "^1.3.2",
     "@react-navigation/bottom-tabs": "^6.5.11",
+    "@react-navigation/material-bottom-tabs": "^6.2.19",
     "@react-navigation/native": "^6.1.9",
     "@react-navigation/native-stack": "^6.9.17",
     "@reduxjs/toolkit": "^2.0.1",
@@ -20,7 +21,7 @@
     "react": "18.2.0",
     "react-i18next": "^14.0.1",
     "react-native": "0.72.6",
-    "react-native-paper": "^5.12.1",
+    "react-native-paper": "^5.12.3",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
     "react-native-vector-icons": "^10.0.3",
